engine/requirements.txt
fastapi
uvicorn
docker
gitpython
sqlalchemycryptography
pytest
httpx